Did somebody say ONE BIG CHOMP CHOMP ROCKET
Revamped and finished spawners.

Changed all the binds to be universalized as clockwise AKA up, right, down, left AKA the way I am/was rendering arrows.

Made a new base sprite with moar corners.

Spawners currently have colors, which are entirely unused and just default to 0xFFFFFFFF. Later this could be used for powerups by changing the sprite to be white/grey and making speedy modes be yellow and slow modes be blue and bad modes to be green or something.

You can still currently place arrows on player's bases, so I'm going to revamp bases tomorrow and get started on spawning and rendering the pups. Maybe I'll create a debug bind for create dog or something, so we can worry about how often dogs should be spawned later.