I BUILT A WALL AND THEY'RE GOING TO PAY FOR IT
Walls exist. Creatures run into the sides of the level and into walls. Currently, the walls seem to have a drawing problem or possibly a collision thing. I'll look into it soon.

Creatures need to collide with arrows still. I also will have to change the behavior of turning when you run into stuff.

In the current chomp chomp, you just turn right theoretically until you can move again. (there seems to be a teleporting bug thing that I will test tomorrow and fix maybe)
In choo choo, you turn right and if there's a wall there you turn left instead, and if there's a wall there you turn backwards. So ours is different in that it doesn't turn left if there's a second collision immediately, it skips straight to backwards with a right right.

Tomorrow stuff:
Creatures colliding with arrows
Remove the weird sidestep teleport thing with multiple collisions
Possibly change the turning behavior, or make that a setting. Actually, yeah make that a setting.